# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API",
        "version": "1.0"
    },
    "paths": {
...
        "/dashboard/background": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Récupère tous les groupes et arrière-plans disponibles.",
                "description": "Récupère tous les groupes et arrière-plans disponibles.",
                "operationId": "49c538684827e4255048a81a9c35c5e2",
                "responses": {
                    "200": {
                        "description": "Liste des groupes et arrière-plans disponibles."
                    },
                    "404": {
                        "description": "Groupe d'arrière-plan introuvable."
                    }
                }
            }
        },
        "/dashboard/background/{id}": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Met à jour l'arrière-plan d'un livret avec l'arrière-plan sélectionné.",
                "description": "Met à jour l'arrière-plan d'un livret avec l'arrière-plan sélectionné.",
                "operationId": "e96d053f9a5c952f3a967c246234770d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arrière-plan mis à jour avec succès."
                    },
                    "404": {
                        "description": "Arrière-plan introuvable."
                    }
                }
            }
        },
...
    },
...
}